Separate vconsole and livereload plugin
