+++ .eslintrc.json
@@ -18,7 +18,8 @@
         ],
         "quotes": [
             "error",
-            "single"
+            "single",
+            { "allowTemplateLiterals": true }
         ],
         "semi": [
             "error",
