+++ packages/foxman-plugin-livereload/lib/reloader.js
