+++ packages/foxman-plugin-vconsole/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "@foxman/plugin-vconsole",
+  "version": "0.1.0",
+  "description": "",
+  "main": "lib/index.js",
+  "scripts": {
+  },
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+  }
+}
